requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
google-generativeai>=0.4
pandas>=2.2

# Optional (only used when SHEETS_MODE=google)
google-api-python-client>=2.120
google-auth>=2.28
